requirements.txt
cycler==0.11.0
dataclasses==0.8
future==0.18.2
keyboard==0.13.5
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.5.62
Pillow==8.4.0
protobuf==3.19.4
pygame==2.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
scipy==1.5.4
six==1.16.0
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
typing_extensions==4.1.1
